requirements.txt
crispy-tailwind==0.5.0
Django==4.1.3
django-crispy-forms==1.14.0
djangorestframework==3.14.0
